refactor: 페르소나 의견 수집 시작 시 사용자 크레딧 차감 (#82)

+++ src/main/java/kr/yuns/dropthepitchserver/credit/data/repository/FileAnalysisCreditRepository.java
@@ -3,5 +3,5 @@ package kr.yuns.dropthepitchserver.credit.data.repository;
 import kr.yuns.dropthepitchserver.credit.data.entity.FileAnalysisCredit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FileAnalysisCreditRepository extends JpaRepository<FileAnalysisCredit, Integer> {
+public interface FileAnalysisCreditRepository extends JpaRepository<FileAnalysisCredit, Long> {
 }
